requirements.txt
aioredis==1.2.0
aniso8601==3.0.2
asgiref==2.3.2
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
backcall==0.1.0
channels==2.2.0
channels-redis==2.3.2
constantly==15.1.0
daphne==2.3.0
decorator==4.4.0
Django==2.1.7
django-extra-fields==1.2.1
djangorestframework==3.8.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
msgpack==0.5.6
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
Pygments==2.3.1
PyHamcrest==1.9.0
python-decouple==3.1
pytz==2018.9
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
traitlets==4.3.2
Twisted==19.2.0
txaio==18.8.1
wcwidth==0.1.7
zope.interface==4.6.0
